requirements.txt
aiofiles==25.1.0
aiogram==3.27.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
asyncpg==0.31.0
attrs==26.1.0
certifi==2026.2.25
dotenv==0.9.9
frozenlist==1.8.0
greenlet==3.4.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.1
pillow==12.2.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
PyYAML==6.0.3
qrcode==8.2
redis==7.4.0
SQLAlchemy==2.0.49
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.23.0
